feat: 메서드 합치기 #83
