+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/data/repository/AuthRepository.kt
@@ -4,21 +4,22 @@ import android.content.Context
 import android.util.Log
 import com.wonchihyeon.ytt_android.data.model.ResponseDTO
 import com.wonchihyeon.ytt_android.data.model.SignInDTO
+import com.wonchihyeon.ytt_android.data.model.SignUpDTO
 import com.wonchihyeon.ytt_android.data.network.ApiService
 import com.wonchihyeon.ytt_android.data.network.RetrofitAPI
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class SignInRepository(private val context: Context) {
-
-    private val apiService: ApiService = RetrofitAPI.getAuthRetrofit(context).create(ApiService::class.java)
+class AuthRepository(private val context: Context) {
+    private val apiService: ApiService =
+        RetrofitAPI.getAuthRetrofit(context).create(ApiService::class.java)
 
     fun signIn(signInDTO: SignInDTO, callback: (String, String?, String?) -> Unit) {
         apiService.signIn(signInDTO).enqueue(object : Callback<ResponseDTO<String>> {
             override fun onResponse(
                 call: Call<ResponseDTO<String>>,
-                response: Response<ResponseDTO<String>>
+                response: Response<ResponseDTO<String>>,
             ) {
                 if (response.isSuccessful) {
                     // 응답 헤더 전체 출력
@@ -39,6 +40,30 @@ class SignInRepository(private val context: Context) {
             override fun onFailure(call: Call<ResponseDTO<String>>, t: Throwable) {
                 callback("로그인 실패: ${t.message}", null, null)
             }
+        })
+    }
+
+    fun signUp(signUpDTO: SignUpDTO, callback: (String, String?, String?) -> Unit) {
+        apiService.signUp(signUpDTO).enqueue(object : Callback<ResponseDTO<String>> {
+            override fun onResponse(
+                call: Call<ResponseDTO<String>>,
+                response: Response<ResponseDTO<String>>,
+            ) {
+                Log.d("response_code", response.code().toString())
+                if (response.isSuccessful) {
+                    val responseBody: String = response.body().toString()
+                    callback("회원가입 성공", responseBody, null)
+                } else {
+                    val errorMessage = response.errorBody()?.string() ?: "알 수 없는 오류"
+                    callback("회원가입 실패: $errorMessage", null, null)
+                }
+            }
+
+            override fun onFailure(call: Call<ResponseDTO<String>>, t: Throwable) {
+                Log.e("SignUpRepository", "회원가입 요청 실패: ${t.message}", t)
+                callback("회원가입 요청 실패: ${t.message}", null, null)
+            }
+
         })
     }
 }
